AirportItlwm: add macOS Tahoe (26.0) support
Add IO80211ControllerV3 header and AirportItlwm-Tahoe build target.
Adapt Skywalk interface, controller, and HAL code for Tahoe vtable
changes: new init signature, moved setLinkState, removed enable/disable
and apple80211_ioctl methods, const getWorkQueue, new IOCTLs, and
guard __PRIVATE_SPI__ output thread calls.

+++ itlwm/hal_iwm/nvm.cpp
@@ -160,7 +160,7 @@ iwm_nvm_read_chunk(struct iwm_softc *sc, uint16_t section, uint16_t offset,
     struct iwm_host_cmd cmd = {
         .id = IWM_NVM_ACCESS_CMD,
         .flags = (IWM_CMD_WANT_RESP | IWM_CMD_SEND_IN_RFKILL),
-        .resp_pkt_len = IWM_CMD_RESP_MAX,
+        .resp_pkt_len = (size_t)IWM_CMD_RESP_MAX,
         .data = { &nvm_access_cmd, },
     };
     int err, offset_read;
